com.unity.test-framework@1.4.3 ## [1.4.3] - 2024-01-15 - Fixed a performance issue in relation to the new ui, where running a large selection of tests did not perform well. (DSTR-983) - The test tree no longer is fully expanded when the window is first opened, fixing test tree explosion. (DSTR-985) - Fixed an issue in Solution 13 of Samples where the path was pointing to a wrong location. (DSTR-1005) - Updated documentation of Solution 13 of Samples to reflect more accurate solution for the example. (DSTR-1005) - Added ability to build all or selected tests from test runner UI. (UTF-553)

+++ Samples~/13_DomainReload_Solution/Tests_13s/ScriptAddingTests.cs
@@ -9,7 +9,7 @@ namespace Tests_13s
 {
     internal class ScriptAddingTests
     {
-        private const string pathToFile = "Assets/Tests/TempScript.cs"; 
+        private const string pathToFile = "Assets/Tests_13s/TempScript.cs"; 
         
         [UnityTest]
         public IEnumerator CreatedScriptIsVerified()
@@ -29,8 +29,13 @@ namespace Tests_13s
             {
                 yield break;
             }
+            if (!File.Exists(pathToFile + ".meta"))
+            {
+                yield break;
+            }
             
             File.Delete(pathToFile);
+            File.Delete(pathToFile + ".meta");
             yield return new RecompileScripts();
         }
